+++ src/main/scala/com/databricks/labs/automl/pipeline/inference/PipelineModelInference.scala
@@ -2,7 +2,7 @@ package com.databricks.labs.automl.pipeline.inference
 
 import com.databricks.labs.automl.executor.config.LoggingConfig
 import com.databricks.labs.automl.params.{MLFlowConfig, MainConfig}
-import com.databricks.labs.automl.utils.AutoMlPipelineMlFlowUtils
+import com.databricks.labs.automl.utils.{AutoMlPipelineMlFlowUtils, InitDbUtils}
 import org.apache.spark.ml.PipelineModel
 
 /**
@@ -12,14 +12,37 @@ import org.apache.spark.ml.PipelineModel
   */
 object PipelineModelInference {
 
+
   /**
-    * Run Inference directly against a given MlFlow Run ID
+    *
+    * @param runId String of MLFlow runId to be used for Inference
+    * @param loggingConfig Deprecated -- logging config for older pipelines
+    * @return
+    */
+  @deprecated("Only for legacy pipelines without main config tracked by MLFlow. Use " +
+    "signature (runId: String, mainConfig: mainConfig: MainConfig) or " +
+    "(runId: String)", "0.7.1")
+  def getPipelineModelByMlFlowRunId(runId: String, loggingConfig: LoggingConfig): PipelineModel = {
+    PipelineModel.load(AutoMlPipelineMlFlowUtils.getPipelinePathByRunId(runId, loggingConfig=Some(loggingConfig)))
+  }
+
+  /***
+    * String of MLFlow runId to be used for Inference
     * @param runId
     * @param mainConfig
     * @return
     */
   def getPipelineModelByMlFlowRunId(runId: String, mainConfig: MainConfig): PipelineModel = {
-    PipelineModel.load(AutoMlPipelineMlFlowUtils.getPipelinePathByRunId(runId, mainConfig))
+    PipelineModel.load(AutoMlPipelineMlFlowUtils.getPipelinePathByRunId(runId, mainConfig=Some(mainConfig)))
+  }
+
+  /**
+    * String of MLFlow runId to be used for Inference
+    * @param runId
+    * @return
+    */
+  def getPipelineModelByMlFlowRunId(runId: String): PipelineModel = {
+    PipelineModel.load(AutoMlPipelineMlFlowUtils.getPipelinePathByRunId(runId, None))
   }
 
 }